backend/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "execsos-backend"
version = "2.0.0"
description = "ExecOS AI-powered executive boardroom backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy[asyncio]",
    "asyncpg",
    "alembic",
    "email-validator",
    "crewai[google]",
    "python-dotenv",
    "sse-starlette",
    "pydantic[email]",
    "pydantic-settings",
    "passlib[bcrypt]",
    "python-jose[cryptography]",
    "mem0ai",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
